Now gen_source_docs handles thread object types
Lasso shows a thread object type as two types: one that has the name it was
given and one that has the name with "_thread$" tagged on to the end. While the
member methods are all under the given name, the docComment is stored in the one
that adds "_thread$" to it.

The code now handles this by copying the docComment to the named thread type.
Since Lasso defines the named thread type after the one that adds "_thread$",
by the time the named thread type is processed, it has the documentation from
the docComment of the other.